- Made sure all unit tests always wait for any indexing before querying.   (ie ConsistencyOptions.QueryYourWrites). - 2nd Unit test is now working great :) Added ordering into the Index it uses for Popular Tags.

+++ Code/RavenOverflow.Web/Models/ViewModels/IndexViewModel.cs
@@ -12,7 +12,7 @@ namespace RavenOverflow.Web.Models.ViewModels
 
         public IList<Question> Questions { get; set; }
         public AuthenticationViewModel AuthenticationViewModel { get; set; }
-        public IDictionary<string, int> RecentPopularTags { get; set; }
+        public IDictionary<string, short> RecentPopularTags { get; set; }
         public UserTagListViewModel UserFavoriteTagList { get; set; }
         public UserTagListViewModel UserIgnoredTagList { get; set; }
     }
